+++ src/main/scala/edu/illinois/ncsa/fence/models/Stats.scala
@@ -3,6 +3,6 @@ package edu.illinois.ncsa.fence.models
 import com.fasterxml.jackson.annotation.JsonProperty
 
 /** Statistics based on activity */
-case class Stats(bytes: BytesStats, conversionsNum: Int, extractionsNum: Int, keysNum: Int, tokensNum: Int)
+case class Stats(bytes: BytesStats, conversionsNum: Long, extractionsNum: Long, keysNum: Long, tokensNum: Long)
 
-case class BytesStats(conversions_bytes: Int, @JsonProperty("extraction_bytes") extractionBytes: Int)
+case class BytesStats(conversions_bytes: Long, @JsonProperty("extraction_bytes") extractionBytes: Long)
